refactor: modifying architecure to fits the clean architecture
