+++ src/main/java/com/example/SpringChat/entity/AbstractEntity.java
@@ -1,4 +1,4 @@
-package com.example.SpringChat.model;
+package com.example.SpringChat.entity;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
@@ -11,7 +11,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.UpdateTimestamp;
 import org.hibernate.annotations.UuidGenerator;
