+++ src/main/java/com/example/SpringChat/entity/user/gateway/FriendshipGatewayPort.java
@@ -1,0 +1,6 @@
+package com.example.SpringChat.entity.user.gateway;
+import java.util.List;
+
+public interface FriendshipGatewayPort {
+    List<User> findFriendsByUserId(String userId);
+}
