+++ src/main/java/com/example/SpringChat/entity/user/gateway/UserGatewayPort.java
@@ -1,0 +1,14 @@
+package com.example.SpringChat.entity.user.gateway;
+
+import com.example.SpringChat.entity.user.model.User;
+
+import java.util.Optional;
+
+public interface UserGatewayPort {
+
+    boolean UserEmailExists(String email);
+
+    User save(User user);
+
+    Optional<User> findByEmail(String email);
+}
